refactor(directive): code-spiltting
