+++ package.json
@@ -50,10 +50,10 @@
   "module": "dist/index.mjs",
   "types": "index.d.ts",
   "scripts": {
-    "build": "rimraf dist && tsup src/*.ts --format cjs,esm --dts --splitting && esno scripts/postbuild.ts",
-    "dev": "tsup src/*.ts --format cjs,esm --watch src",
-    "example:build": "npm -C examples/vue3 run build",
-    "example:dev": "npm -C examples/vue3 run dev",
+    "build": "tsup src/*.ts && esno scripts/postbuild.ts",
+    "dev": "tsup src/*.ts --watch src",
+    "example:build": "npm -C examples/vite-vue3 run build",
+    "example:dev": "npm -C examples/vite-vue3 run dev",
     "prepublishOnly": "npm run build",
     "release": "npx bumpp --commit --tag --push",
     "test": "jest",
