+++ src/core/transforms/directive/vue2.ts
@@ -1,16 +1,10 @@
-import Debug from 'debug'
 import type {
   CallExpression, ObjectProperty, File, VariableDeclaration, FunctionExpression, BlockStatement,
 } from '@babel/types'
 import type MagicString from 'magic-string'
 import { parse, ParseResult } from '@babel/parser'
 import traverse from '@babel/traverse'
-import { pascalCase, stringifyComponentImport } from '../utils'
-import type { Context } from '../context'
-import { ResolveResult } from '../transformer'
-import { SupportedTransformer } from '../..'
-
-const debug = Debug('unplugin-vue-components:transform:directive')
+import { ResolveResult } from '../../transformer'
 
 /**
  * get Vue 2 render function position
@@ -20,8 +14,8 @@ const debug = Debug('unplugin-vue-components:transform:directive')
 const getRenderFnStart = (ast: ParseResult<File>): number => {
   const renderFn = ast.program.body.find((node): node is VariableDeclaration =>
     node.type === 'VariableDeclaration'
-    && node.declarations[0].id.type === 'Identifier'
-    && node.declarations[0].id.name === 'render',
+      && node.declarations[0].id.type === 'Identifier'
+      && node.declarations[0].id.name === 'render',
   )
   const start = (((renderFn?.declarations[0].init as FunctionExpression).body) as BlockStatement).start
   if (start === null)
@@ -29,7 +23,7 @@ const getRenderFnStart = (ast: ParseResult<File>): number => {
   return start + 1
 }
 
-const resolveVue2 = (code: string, s: MagicString): ResolveResult[] => {
+export const resolve = (code: string, s: MagicString): ResolveResult[] => {
   const ast = parse(code, {
     sourceType: 'module',
   })
@@ -51,8 +45,8 @@ const resolveVue2 = (code: string, s: MagicString): ResolveResult[] => {
     const directives = args[1].properties.find(
       (property): property is ObjectProperty =>
         property.type === 'ObjectProperty'
-        && property.key.type === 'Identifier'
-        && property.key.name === 'directives',
+          && property.key.type === 'Identifier'
+          && property.key.name === 'directives',
     )?.value
     if (!directives || directives.type !== 'ArrayExpression')
       continue
@@ -65,8 +59,8 @@ const resolveVue2 = (code: string, s: MagicString): ResolveResult[] => {
       const nameNode = directive.properties.find(
         (p): p is ObjectProperty =>
           p.type === 'ObjectProperty'
-          && p.key.type === 'Identifier'
-          && p.key.name === 'name',
+            && p.key.type === 'Identifier'
+            && p.key.name === 'name',
       )?.value
       if (nameNode?.type !== 'StringLiteral') continue
       const name = nameNode.value
@@ -81,43 +75,4 @@ const resolveVue2 = (code: string, s: MagicString): ResolveResult[] => {
   }
 
   return results
-}
-
-const resolveVue3 = (code: string, s: MagicString): ResolveResult[] => {
-  const results: ResolveResult[] = []
-
-  for (const match of code.matchAll(/_resolveDirective\("(.+?)"\)/g)) {
-    const matchedName = match[1]
-    if (match.index != null && matchedName && !matchedName.startsWith('_')) {
-      const start = match.index
-      const end = start + match[0].length
-      results.push({
-        rawName: matchedName,
-        replace: resolved => s.overwrite(start, end, resolved),
-      })
-    }
-  }
-
-  return results
-}
-
-export default async(code: string, transformer: SupportedTransformer, s: MagicString, ctx: Context, sfcPath: string) => {
-  let no = 0
-
-  const results = transformer === 'vue2' ? resolveVue2(code, s) : resolveVue3(code, s)
-  for (const { rawName, replace } of results) {
-    debug(`| ${rawName}`)
-    const name = pascalCase(rawName)
-    ctx.updateUsageMap(sfcPath, [name])
-
-    const directive = await ctx.findComponent(name, 'directive', [sfcPath])
-    if (!directive) continue
-
-    const varName = `__unplugin_directives_${no}`
-    s.prepend(`${stringifyComponentImport({ ...directive, name: varName }, ctx)};\n`)
-    no += 1
-    replace(varName)
-  }
-
-  debug(`^ (${no})`)
 }
