+++ src/core/transforms/directive/vue3.ts
@@ -1,0 +1,20 @@
+import type MagicString from 'magic-string'
+import { ResolveResult } from '../../transformer'
+
+export const resolve = (code: string, s: MagicString): ResolveResult[] => {
+  const results: ResolveResult[] = []
+
+  for (const match of code.matchAll(/_resolveDirective\("(.+?)"\)/g)) {
+    const matchedName = match[1]
+    if (match.index != null && matchedName && !matchedName.startsWith('_')) {
+      const start = match.index
+      const end = start + match[0].length
+      results.push({
+        rawName: matchedName,
+        replace: resolved => s.overwrite(start, end, resolved),
+      })
+    }
+  }
+
+  return results
+}
